added '--no-package' option

+++ build.sh
@@ -25,10 +25,12 @@ do_build() {
     if [ -f "dist/${BOX_NAME}.box" ]; then
         rm dist/${BOX_NAME}.box
     fi
-    vagrant package ${BOX_NAME} --output dist/${BOX_NAME}.box
-    vagrant box add src/${BOX_NAME}.json
-    rm dist/${BOX_NAME}.box
-    vagrant destroy ${BOX_NAME} -f
+    if [ $2 == "--no-package" ]; then
+        vagrant package ${BOX_NAME} --output dist/${BOX_NAME}.box
+        vagrant box add src/${BOX_NAME}.json
+        rm dist/${BOX_NAME}.box
+        vagrant destroy ${BOX_NAME} -f
+    fi
 }
 
 do_help() {
@@ -41,6 +43,8 @@ do_help() {
     echo "    precise        Building 'precise' x64 vagrant box"
     echo "    trusty         Building 'trusty' x64 vagrant box"
     echo "    trusty_php7    Building 'trusty' x64 with PHP v7 vagrant box"
+    echo ""
+    echo "    --no-package   only 'vagrant up' without packaging"
     echo ""
 }
